backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
openai>=1.12.0
python-dotenv==1.0.0
pydantic==2.5.0
